Refactor net tests to reflect new module structure.
